+++ demo-cassandra/demo-order-management-app/common/src/main/java/com/future/demo/dto/OrderDetailDTO.java
@@ -2,12 +2,10 @@ package com.future.demo.dto;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 public class OrderDetailDTO {
     private Long id;
-    private BigDecimal orderId;
+    private Long orderId;
     private Long productId;
     private Long merchantId;
     private Integer amount;
