+++ demo-cassandra/demo-order-management-app/common/src/main/java/com/future/demo/entity/OrderDetailModel.java
@@ -2,12 +2,10 @@ package com.future.demo.entity;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 public class OrderDetailModel {
     private Long id;
-    private BigDecimal orderId;
+    private Long orderId;
     private Long userId;
     private Long productId;
     private Long merchantId;
