+++ demo-cassandra/demo-order-management-app/common/src/main/java/com/future/demo/entity/OrderModel.java
@@ -2,14 +2,12 @@ package com.future.demo.entity;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
 @Data
 public class OrderModel {
-    private BigDecimal id;
-
+    private Long id;
     private Long userId;
     private LocalDateTime createTime;
     private Status status;
